Move transformers into their own folder
