+++ packages/core/package.json
@@ -42,25 +42,19 @@
       "Const",
       "Dictionary",
       "Either",
-      "EitherT",
       "FreeAssociative",
       "Hash",
       "Id",
-      "IndexedT",
-      "InvariantT",
       "IO",
       "Iterable",
       "Map",
       "NonEmptyArray",
       "Option",
-      "OptionT",
       "Ordering",
       "Reader",
-      "ReaderT",
       "RedBlackTree",
       "RoseTree",
       "Set",
-      "StateT",
       "SortedSet",
       "String",
       "Effect",
@@ -165,6 +159,12 @@
       "Support/MutableQueue",
       "Support/OneShot",
       "Support/RingBuffer",
+      "Transformer/EitherT",
+      "Transformer/IndexedT",
+      "Transformer/InvariantT",
+      "Transformer/OptionT",
+      "Transformer/ReaderT",
+      "Transformer/StateT",
       "Utils",
       "XPure",
       "XPure/XIO",
