+++ packages/core/src/Modules/index.ts
@@ -12,7 +12,6 @@ import type { Layer } from "../Effect/Layer"
 import type { Either } from "../Either"
 import type { FreeAssociative } from "../FreeAssociative"
 import type { Id } from "../Id"
-import type { Ix, IxC } from "../IndexedT"
 import type { IO } from "../IO"
 import type { List } from "../List"
 import type { NonEmptyArray } from "../NonEmptyArray"
@@ -23,12 +22,13 @@ import type { Ord } from "../Prelude/Ord"
 import type { Show } from "../Prelude/Show"
 import type { Reader } from "../Reader"
 import type { Tree } from "../RoseTree"
-import type { StateIn, StateOut } from "../StateT/Classic"
 import type { Closure } from "../Structure/Closure"
 import type { Commutative } from "../Structure/Commutative"
 import type { Identity } from "../Structure/Identity"
 import type { Inverse } from "../Structure/Inverse"
 import type { Sync } from "../Sync"
+import type { Ix, IxC } from "../Transformer/IndexedT"
+import type { StateIn, StateOut } from "../Transformer/StateT/Classic"
 import type { XPure } from "../XPure"
 import type { XIO } from "../XPure/XIO"
 import type { XReader } from "../XPure/XReader"
