+++ packages/core/src/Transformer/EitherT/index.ts
@@ -1,13 +1,13 @@
 // tracing: off
 
-import "../Operator"
+import "../../Operator"
 
-import * as E from "../Either"
-import { pipe } from "../Function"
-import type { Any, Applicative, Covariant, Monad, URI } from "../Prelude"
-import { succeedF } from "../Prelude/DSL"
-import type { Access, Fail, Provide, Run } from "../Prelude/FX"
-import * as HKT from "../Prelude/HKT"
+import * as E from "../../Either"
+import { pipe } from "../../Function"
+import type { Any, Applicative, Covariant, Monad, URI } from "../../Prelude"
+import { succeedF } from "../../Prelude/DSL"
+import type { Access, Fail, Provide, Run } from "../../Prelude/FX"
+import * as HKT from "../../Prelude/HKT"
 
 export type V<C> = HKT.CleanParam<C, "E"> & HKT.V<"E", "+">
