+++ packages/core/src/Transformer/IndexedT/index.ts
@@ -1,12 +1,12 @@
 // tracing: off
 
-import "../Operator"
+import "../../Operator"
 
-import { pipe } from "../Function"
-import type { IxURI } from "../Modules"
-import { chainF, succeedF } from "../Prelude/DSL"
-import * as HKT from "../Prelude/HKT"
-import type { Monad } from "../Prelude/Monad"
+import { pipe } from "../../Function"
+import type { IxURI } from "../../Modules"
+import { chainF, succeedF } from "../../Prelude/DSL"
+import * as HKT from "../../Prelude/HKT"
+import type { Monad } from "../../Prelude/Monad"
 
 export interface IxC<I, O> {
   Ix: {
