+++ packages/core/src/Transformer/ReaderT/index.ts
@@ -1,13 +1,13 @@
 // tracing: off
 
-import "../Operator"
+import "../../Operator"
 
-import { pipe } from "../Function"
-import type { Applicative, AssociativeEither, Monad } from "../Prelude"
-import { succeedF } from "../Prelude/DSL"
-import type { Access, Fail, Provide, Run } from "../Prelude/FX"
-import * as HKT from "../Prelude/HKT"
-import * as R from "../Reader"
+import { pipe } from "../../Function"
+import type { Applicative, AssociativeEither, Monad } from "../../Prelude"
+import { succeedF } from "../../Prelude/DSL"
+import type { Access, Fail, Provide, Run } from "../../Prelude/FX"
+import * as HKT from "../../Prelude/HKT"
+import * as R from "../../Reader"
 
 export type V<C> = HKT.CleanParam<C, "R"> & HKT.V<"R", "-">
