+++ packages/core/src/Transformer/StateT/Parametric/index.ts
@@ -1,8 +1,8 @@
 // tracing: off
 
-import type { StateInURI, StateOutURI } from "../../Modules"
-import type { Monad } from "../../Prelude"
-import * as HKT from "../../Prelude/HKT"
+import type { StateInURI, StateOutURI } from "../../../Modules"
+import type { Monad } from "../../../Prelude"
+import * as HKT from "../../../Prelude/HKT"
 import * as StateT from "../Classic"
 
 /**
