+++ packages/core/src/Transformer/StateT/index.ts
@@ -1,5 +1,5 @@
 // tracing: off
 
-import "../Operator"
+import "../../Operator"
 
 export * from "./Classic"
