+++ packages/core/src/index.ts
@@ -36,7 +36,7 @@ export * as Take from "./Effect/Stream/Take"
 export * as Transducer from "./Effect/Stream/Transducer"
 export * as Supervisor from "./Effect/Supervisor"
 export * as Either from "./Either"
-export * as EitherT from "./EitherT"
+export * as EitherT from "./Transformer/EitherT"
 export * as Equal from "./Prelude/Equal"
 export * as FreeAssociative from "./FreeAssociative"
 export * as Function from "./Function"
@@ -47,8 +47,8 @@ export * as HashSet from "./HashSet"
 export * as IO from "./IO"
 export * as Id from "./Id"
 export * as Identity from "./Structure/Identity"
-export * as IndexedT from "./IndexedT"
-export * as InvariantT from "./InvariantT"
+export * as IndexedT from "./Transformer/IndexedT"
+export * as InvariantT from "./Transformer/InvariantT"
 export * as Inverse from "./Structure/Inverse"
 export * as Iterable from "./Iterable"
 export * as List from "./List"
@@ -56,7 +56,7 @@ export * as Map from "./Map"
 export * as Newtype from "./Newtype"
 export * as NonEmptyArray from "./NonEmptyArray"
 export * as Option from "./Option"
-export * as OptionT from "./OptionT"
+export * as OptionT from "./Transformer/OptionT"
 export * as Ord from "./Prelude/Ord"
 export * as Ordering from "./Ordering"
 export * as Any from "./Prelude/Any"
@@ -116,15 +116,15 @@ export * as WiltableWithIndex from "./Prelude/WiltableWithIndex"
 export * as Witherable from "./Prelude/Witherable"
 export * as WitherableWithIndex from "./Prelude/WitherableWithIndex"
 export * as Reader from "./Reader"
-export * as ReaderT from "./ReaderT"
+export * as ReaderT from "./Transformer/ReaderT"
 export * as Record from "./Dictionary"
 export * as RedBlackTree from "./RedBlackTree"
 export * as RoseTree from "./RoseTree"
 export * as Set from "./Set"
 export * as Show from "./Prelude/Show"
 export * as SortedSet from "./SortedSet"
-export * as StateT from "./StateT"
-export * as ParametricStateT from "./StateT/Parametric"
+export * as StateT from "./Transformer/StateT"
+export * as ParametricStateT from "./Transformer/StateT/Parametric"
 export * as String from "./String"
 export * as AtomicBoolean from "./Support/AtomicBoolean"
 export * as AtomicNumber from "./Support/AtomicNumber"
