+++ packages/core/test/examples/10.test.ts
@@ -2,8 +2,8 @@ import { pipe } from "@effect-ts/system/Function"
 
 import * as A from "../../src/Array"
 import * as E from "../../src/Either"
-import * as EitherT from "../../src/EitherT"
 import { chainF } from "../../src/Prelude/DSL"
+import * as EitherT from "../../src/Transformer/EitherT"
 import * as R from "../../src/XPure/XReader"
 import * as ReaderT from "../../src/XPure/XReaderT"
