+++ packages/core/test/examples/12.test.ts
@@ -2,9 +2,9 @@ import { pipe } from "@effect-ts/system/Function"
 
 import * as A from "../../src/Array"
 import * as E from "../../src/Either"
-import * as EitherT from "../../src/EitherT"
-import * as InvariantT from "../../src/InvariantT"
 import { chainF } from "../../src/Prelude/DSL"
+import * as EitherT from "../../src/Transformer/EitherT"
+import * as InvariantT from "../../src/Transformer/InvariantT"
 import * as R from "../../src/XPure/XReader"
 import * as ReaderT from "../../src/XPure/XReaderT"
