+++ packages/core/test/examples/13.test.ts
@@ -7,12 +7,12 @@ import { NoSuchElementException } from "@effect-ts/system/GlobalExceptions"
 import { isEither, isOption } from "@effect-ts/system/Utils"
 
 import * as Either from "../../src/Either"
-import * as EitherT from "../../src/EitherT"
 import * as Option from "../../src/Option"
 import { GenHKT } from "../../src/Prelude/DSL"
 import * as DSL from "../../src/Prelude/DSL"
 import * as S from "../../src/Prelude/Selective"
 import { makeAssociative } from "../../src/Structure/Associative"
+import * as EitherT from "../../src/Transformer/EitherT"
 import type { XIO } from "../../src/XPure/XIO"
 import * as IO from "../../src/XPure/XIO"
 import * as Reader from "../../src/XPure/XReader"
