+++ packages/core/test/examples/17.test.ts
@@ -2,10 +2,10 @@ import { flow, identity, pipe } from "@effect-ts/system/Function"
 import * as M from "@effect-ts/system/Map"
 
 import * as E from "../../src/Either"
-import * as EitherT from "../../src/EitherT"
 import * as P from "../../src/Prelude"
 import * as DSL from "../../src/Prelude/DSL"
 import type * as H from "../../src/Prelude/HKT"
+import * as EitherT from "../../src/Transformer/EitherT"
 import * as T from "../../src/XPure"
 import * as R from "../../src/XPure/XReader"
 import * as ReaderT from "../../src/XPure/XReaderT"
